llm-chatbot/requirements.txt
# Core dependencies
openai>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
pandas>=1.5.0
openpyxl>=3.0.0

# Google Drive integration (optional)
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# Web application
Flask>=2.3.0
Werkzeug>=2.3.0

# Environment and utilities
python-dotenv>=1.0.0
numpy>=1.24.0
pathlib2>=2.3.0

# Development and testing
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0